+++ app/src/main/java/com/wildplot/android/rendering/LinesPoints.java
@@ -41,9 +41,9 @@ public class LinesPoints implements Drawable {
 		Color oldColor = g.getColor();
 		Rectangle field = g.getClipBounds();
 		g.setColor(color);
-		
-		int[] coordStart = plotSheet.toGraphicPoint(pointList[0][0],pointList[1][0],field);
-		int[] coordEnd = coordStart;
+
+        float[] coordStart = plotSheet.toGraphicPoint(pointList[0][0],pointList[1][0],field);
+        float[] coordEnd = coordStart;
 		
 		for(int i = 0; i< pointList[0].length; i++) {
 			coordEnd = coordStart;
@@ -62,7 +62,7 @@ public class LinesPoints implements Drawable {
 	 * @param field given rectangle field
 	 */
 	public void drawPoint(double x, double y, Graphics g, Rectangle field) {
-		int[] coordStart 	= plotSheet.toGraphicPoint(x, y,field);
+        float[] coordStart 	= plotSheet.toGraphicPoint(x, y,field);
 		g.drawRect(coordStart[0]-3, coordStart[1]-3, 6, 6);
 //		g.drawLine(coordStart[0]-3, coordStart[1]-3, coordStart[0]+3, coordStart[1]-3);
 //		g.drawLine(coordStart[0]+3, coordStart[1]-3, coordStart[0]+3, coordStart[1]+3);
