+++ app/src/main/java/com/wildplot/android/rendering/graphics/wrapper/Graphics.java
@@ -24,21 +24,21 @@ public class Graphics {
         this.paint = paint;
     }
 
-    public void drawLine(int x1, int y1, int x2, int y2){
+    public void drawLine(float x1, float y1, float x2, float y2){
         Style oldStyle = paint.getStyle();
         paint.setStyle(Style.STROKE);
         canvas.drawLine(x1, y1, x2, y2, paint);
         paint.setStyle(oldStyle);
     }
     
-    public void drawRect(int x, int y, int width, int height){
+    public void drawRect(float x, float y, float width, float height){
         Style oldStyle = paint.getStyle();
         paint.setStyle(Style.STROKE);
         canvas.drawRect(x, y, x+width, y+height, paint);
         paint.setStyle(oldStyle);
     }
     
-    public void fillRect(int x, int y, int width, int height){
+    public void fillRect(float x, float y, float width, float height){
         Style oldStyle = paint.getStyle();
         paint.setStyle(Style.FILL_AND_STROKE);
         canvas.drawRect(x, y, x+width, y+height, paint);
@@ -69,20 +69,20 @@ public class Graphics {
         paint.setColor(color.getColorValue());
     }
     
-    public void drawArc(int x, int y, int width, int height, int startAngle, int arcAngle){
+    public void drawArc(float x, float y, float width, float height, float startAngle, float arcAngle){
         Style oldStyle = paint.getStyle();
         paint.setStyle(Style.STROKE);
         canvas.drawCircle(x, y, width/2.0f, paint);
         paint.setStyle(oldStyle);
     }
-    public void fillArc(int x, int y, int width, int height, int startAngle, int arcAngle){
+    public void fillArc(float x, float y, float width, float height, float startAngle, float arcAngle){
         Style oldStyle = paint.getStyle();
         paint.setStyle(Style.FILL_AND_STROKE);
         canvas.drawCircle(x, y, width/2.0f, paint);
         paint.setStyle(oldStyle);
     }
     
-    public void drawImage(BufferedImage image, String tmp, int x, int y){
+    public void drawImage(BufferedImage image, String tmp, float x, float y){
         //System.err.println("drawImage: " + image.getBitmap().getWidth() + " : "+ image.getBitmap().getHeight());
         Xfermode mode  = paint.getXfermode();
         paint.setXfermode(new PorterDuffXfermode(Mode.SRC_OVER));
@@ -91,7 +91,7 @@ public class Graphics {
         paint.setXfermode(mode);
     }
     
-    public void drawString(String text, int x, int y){
+    public void drawString(String text, float x, float y){
         Style oldStyle = paint.getStyle();
         paint.setStyle(Style.FILL);
         canvas.drawText(text, x, y, paint);
